Move logic to check for users and priviledge of user in courses to new courseAuthService class
(cherry picked from commit f0c188e4395f88d1dc5ef898639c3e3b73d43959)
